synq: clamp formatting TextEdit range to uinteger bounds
The LSP spec defines Position.line as uinteger (0 to 2^31-1).
u32::MAX exceeds this. Fixes #66.

+++ syntaqlite/src/lsp/server.rs
@@ -479,7 +479,7 @@ impl LspServer {
         match host.format(uri, &config) {
             Ok(formatted) => {
                 let edit = TextEdit {
-                    range: Range::new(Position::new(0, 0), Position::new(u32::MAX, 0)),
+                    range: Range::new(Position::new(0, 0), Position::new(i32::MAX as u32, 0)),
                     new_text: formatted,
                 };
                 Response::new_ok(req.id, Some(vec![edit]))
